Change OpenStack to OpenShift
There's a typo in the "Adding destination providers" section: we refer
to OpenStack instead of OpenShift when explaining that providing blank
URL and Service account token in the add-provider dialog means the local
cluster will be used. This PR replaces "local OpenStack cluster" with
"local OpenShift cluster" to fix this.

Signed-off-by: Arik Hadas <ahadas@redhat.com>